pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dark-rl"
version = "0.1.0"
authors = [{ name = "Patrick Barker" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
description = "Advanced Kernels for Reinforcement Learning"
requires-python = ">=3.11,<3.13"
dependencies = [
    "torch>=2.4.0",
    "triton>=3.0.0",
    "transformers>=4.51.0",
    "flash-attn==2.8.0.post2",
    "bitsandbytes>=0.43.0",
    "nvidia-ml-py",
    "xxhash",
    "json-repair>=0.47.1",
    "pillow>=11.2.1",
    "torchvision>=0.22.1",
    "mcp-use>=1.3.3",
]

[project.optional-dependencies]
test = [
    "pytest"
]

[project.urls]
Homepage="https://github.com/pbarker/dark.rl"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["dark"]
